Update ArgumentParser.java
Fixed bug in argument checking

There are supposed to be 9 arguments (args[0] to args[8]). If there are 8 arguments, the if statement does not catch this (this is because 8 < 8 is False), so the program crahses on line 31 when it tries to access args[8].

Doing 8 <= 8 fixes this, as 8 <= 8 is True so the if statement will be able to catch this!